requirements.txt
discord.py
requests
pyNaCl
gTTS
google-cloud-firestore
google-cloud-texttospeech
google-cloud-logging